fix(shhhhh): pixel-hand parity with ShareAssetD3 + spacing fix
Two Phase 0 polish items on the merged /shhhhh landing page:

1. Pixelation now matches our 3D card asset. The hero card was a flat
   pre-pixelated PNG (cart-gradient-10-pixel) that looked rough. Extracted
   <PixelatedCardFace /> from ShareAssetD3 so /shhhhh and the in-app share
   asset use the SAME runtime-pixelated peanut-card-hand SVG. Single source
   of truth for the pixelation algorithm.

2. 'ONLY 20a week' → 'ONLY 20 a week'. Inline-block pill was eating the
   trailing whitespace; added explicit {' '} after the ScarcityCounter
   component.

Also fixes a latent bug in the share-asset cache: cloneNode(true) on a
canvas does not copy the drawing surface, so every 'cached' mount after
the first was blank. Now copies via drawImage(cachedCanvas, 0, 0).
