+++ src/assets/cards/index.ts
@@ -2,7 +2,6 @@ import CARD_GRADIENT_4 from './Cart Gradient 4.svg'
 import CARD_GRADIENT_5 from './Cart Gradient 5.png'
 import CARD_GRADIENT_9 from './Cart Gradient 9.svg'
 import CARD_GRADIENT_10 from './Cart Gradient 10.svg'
-import CARD_GRADIENT_10_PIXEL from './cart-gradient-10-pixel.png'
 import VISA_BRAND_MARK from './visa-brand-mark.png'
 import PEANUT_CARD_HAND from './peanut-card-hand.svg'
 
@@ -20,7 +19,6 @@ export {
     CARD_GRADIENT_5,
     CARD_GRADIENT_9,
     CARD_GRADIENT_10,
-    CARD_GRADIENT_10_PIXEL,
     VISA_BRAND_MARK,
     PEANUT_CARD_HAND,
 }
